% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC}
\alias{AUC}
\title{AUC function}
\usage{
AUC(ROC)
}
\arguments{
\item{ROC}{ROC object}
}
\description{
This function calculates the area under the curve (AUC)
for a Receiver Operator Characteristic (ROC) object
}
\examples{
AUC(ROC)
}
\keyword{AUC}
