% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCsd.R
\name{AUCsd}
\alias{AUCsd}
\title{AUC function}
\usage{
AUCsd(ROC)
}
\arguments{
\item{ROC}{ROC object}
}
\description{
This function calculates 1-sigma error on the AUC
for a Receiver Operator Characteristic (ROC) object
}
\examples{
AUCsd(ROC)
}
\keyword{1-sigma}
\keyword{AUC,}
\keyword{area,}
\keyword{confidence}
