% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestpamk.R
\name{bestpamk}
\alias{bestpamk}
\title{bestpamk function}
\usage{
bestpamk(dataset)
}
\arguments{
\item{dataset}{matrix or dataframe, rows are instances, columns are features}
}
\description{
This function finds the "best" number of clusters to split a dataset into
based on unsupervised clustering using the pam function in library cluster
}
\examples{
bestpamk(data.frame(A=runif(100),B=runif(100)))
}
\keyword{analysis}
\keyword{cluster}
