% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{completeness}
\alias{completeness}
\title{AUC function}
\usage{
completeness(R2D, R2Dcut, L, Lcut)
}
\arguments{
\item{R2D}{vector of 2D- radial positions of stars}

\item{R2Dcut}{scale radius (e.g. half-mass radius)}

\item{L}{vector of stellar luminosities}

\item{Lcut}{scale luminosity (e.g. 0.5 solar)}
}
\description{
This function returns the probability that a star at rho and L is seen
by using an analytical recipe for completeness
}
\examples{
completeness(r, 1.0, L, 0.5)
}
\keyword{completeness}
