% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceIntervals.R
\name{confi}
\alias{confi}
\title{confi function}
\usage{
confi(j, colo = "black", ncurves = 8, ncROC, ROC_tables, shadedsigma = 2)
}
\description{
This function plots ROC curves and their confidence intervals
}
\keyword{ROC}
