% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{confusion}
\alias{confusion}
\title{confusion function}
\usage{
confusion(ROC)
}
\arguments{
\item{ROC}{ROC object}
}
\description{
This function returns a list with the false positive rate (FPR)
and true positive rates (TPR) for a ROC object. The TPR and FPR
are calculated at the furthest point of the ROC curve from the
diagonal identity line, i.e. at the optimal threshold for the
classifier assuming equal cost for false positives and false negatives
}
\examples{
confusion(ROC)
}
\keyword{FPR,}
\keyword{TPR,}
\keyword{optimal}
\keyword{point}
