% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurePlot.R
\name{feature_plot}
\alias{feature_plot}
\title{function feature_plot}
\usage{
feature_plot(rbin, ds)
}
\arguments{
\item{rbin}{(mid?) radii of bins}

\item{ds}{surface brightness at a given bin}
}
\description{
This function plots features from a surface brightness profile
}
\examples{
feature_plot(rbin, ds)
}
\keyword{brightness}
\keyword{features,}
\keyword{surface}
