% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOVSelect.R
\name{FOVselect}
\alias{FOVselect}
\title{FOVselect function}
\usage{
FOVselect(x, y, xi_FOV, alpha_FOV, beta_FOV, rh)
}
\arguments{
\item{x}{vector with x positions of all stars}

\item{y}{vector with y positions of all stars}

\item{xi_FOV}{position of the FOV along the x axis in units of rh (xi_FOV = 0 means centered FOV)}

\item{alpha_FOV}{size of the FOV along the x axis in units of rh}

\item{beta_FOV}{size of the FOV along the y axis in units of rh}

\item{rh}{half-mass radius, used to scale xi_FOV, alpha_FOV, beta_FOV}
}
\description{
This function selects stars that are within the field of view
}
\examples{
FOVselect(x, y, 0, 1, 1, sqrt(median(x*x + y*y))) #square FOV, centered
}
\keyword{FOV}
