% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{impute function}
\usage{
impute(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
This function imputes NAs and Infs with column median
}
\examples{
a <- sample(c(NA, 3, 7), 100, replace = TRUE)
b <- runif(100)
impute(data.frame(a,b))
}
\keyword{Inf}
\keyword{NA,}
\keyword{impute,}
