% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justPlot.R
\name{justplot}
\alias{justplot}
\title{justplot function}
\usage{
justplot(dataset)
}
\arguments{
\item{dataset}{the features+label dataset}
}
\description{
This function plots the snapshots in feature space, showing where the snaps
with a black hole cluster in feature space and where they do not
}
\examples{
justplot(dataset)
}
\keyword{feature}
\keyword{plot}
\keyword{space,}
