% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightn.R
\name{lightn}
\alias{lightn}
\title{lightn function}
\usage{
lightn(colo)
}
\arguments{
\item{colo}{input color}
}
\description{
This function generates a transparent version of the given color
for plotting shaded areas
}
\examples{
lightn("orange") #transparent orange
}
\keyword{area}
\keyword{color,}
\keyword{shaded}
\keyword{transparent,}
