% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ROC_tables.R
\name{make_ROC_tables}
\alias{make_ROC_tables}
\title{make_ROC_tables function}
\usage{
make_ROC_tables()
}
\description{
This function puts ROCs read from ROC files together in a table
}
\examples{
make_ROC_tables()
}
\keyword{ROC}
