% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massDist.R
\name{massDist}
\alias{massDist}
\title{massDist function}
\usage{
massDist()
}
\description{
This function plots the mass distribution of heaviest black holes
}
\examples{
massDist()
}
\keyword{distribution}
\keyword{mass}
