% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pahts_from_list.R
\name{paths_from_list}
\alias{paths_from_list}
\title{function paths_from_list}
\usage{
paths_from_list()
}
\description{
This function reads paths from a file named 'list' with one
path per line. Each path should contain one simulation.
}
\examples{
paths_from_list
}
\keyword{list}
\keyword{path,}
