% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCluster.R
\name{plot_cluster}
\alias{plot_cluster}
\title{plot_cluster function}
\usage{
plot_cluster(x, y, rh = 1, complete, plotfile, incomplete_color = "#888888",
  complete_color = "#000000", FOVbordercolor = "#FF0000", xi_FOV = 0,
  alpha_FOV = 1, beta_FOV = 1)
}
\arguments{
\item{x}{vector with x positions of stars}

\item{y}{vector with y positions of stars}

\item{rh}{scale radius (half mass...)}

\item{complete}{boolean vector that is true for stars that pass completeness (i.e. are included)}

\item{plotfile}{name of the PDF file to plot on}

\item{incomplete_color}{color to plot the excluded stars in (due to completeness)}

\item{complete_color}{color to plot the included stars in}

\item{FOVbordercolor}{color to plot the FOV border}

\item{xi_FOV}{x position of the FOV in units of rh}

\item{alpha_FOV}{x size of the FOV in units of rh}

\item{beta_FOV}{y size of the FOV in units of rh}
}
\description{
This function plots a cluster from x,y star positions
}
\examples{
#plot a cluster with 1000 stars normally distributed in x,y with completeness 0.8
plot_cluster(rnorm(1000),rnorm(1000),1,runif(1000) < 0.8, "cluster.pdf", "#888888", "#000000", "red", 0, 1, 1)
}
\keyword{cluster}
\keyword{plot}
