% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFOV.R
\name{plotFOV}
\alias{plotFOV}
\title{plotFOV function}
\usage{
plotFOV(x_FOV, a_FOV, b_FOV, col)
}
\arguments{
\item{a_FOV}{x size of the FOV in original MOCCA units (i.e. not scaled to rh)}

\item{b_FOV}{y size of the FOV in original MOCCA units (i.e. not scaled to rh)}

\item{col}{color to plot the FOV border}

\item{xi_FOV}{x position of the FOV in original MOCCA units (i.e. not scaled to rh)}
}
\description{
This function plots the FOV as lines on top of a plot
}
\examples{
plotFOV(0, 1, 1, "seagreen") #plots a square, centered FOV in seagreen
}
\keyword{cluster}
\keyword{plot}
