% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsnap.R
\name{plotSnap}
\alias{plotSnap}
\title{plotSnap function}
\usage{
plotSnap(file.pro = "snap.pro", limits = 40 * c(-1, 1))
}
\arguments{
\item{file.pro}{file name of the snapshot, defaults to snap.pro}

\item{limits}{plot limits}
}
\description{
This function plots a snapshot in the xy, xz, yz planes
}
\examples{
plotSnap(file.pro="thissnap") #plots the snapshot in file thissnap (note that the .pro extension is not needed) with default limits
}
\keyword{cluster,}
\keyword{plot}
\keyword{plot,}
\keyword{snap,}
\keyword{snapshot}
\keyword{snapshot,}
