% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\title{profile function}
\usage{
profile(r, m, profbins, feature.plot = FALSE)
}
\arguments{
\item{r}{vector with radial positions of stars}

\item{m}{vector with masses of stars}

\item{profbins}{number of bins}

\item{feature.plot}{if TRUE plots some diagnostic plots}
}
\description{
This function generates a radial density profile given radii and masses
}
\examples{
profile(r, m, 8)
}
\keyword{density}
\keyword{profile}
\keyword{profile,}
\keyword{radial}
