% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeRadii.R
\name{randomize_radii}
\alias{randomize_radii}
\title{randomize_radii function}
\usage{
randomize_radii(r)
}
\arguments{
\item{r}{vector with radial positions of stars}
}
\description{
This function generates x and y positions at random from given 3D radii
}
\examples{
randomize_radii(runif(100)) #generates 100 x y values from random 3D r values
}
\keyword{radii}
\keyword{randomize}
