% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExtendedSnapshot.R
\name{read_ext_snap}
\alias{read_ext_snap}
\title{read_ext_snap function}
\usage{
read_ext_snap(path = "./")
}
\arguments{
\item{path}{path where the snapshot is supposed to be}
}
\description{
This function reads a MOCCA extended snapshot
}
\examples{
read_ext_snap()
}
\keyword{MOCCA}
\keyword{read}
\keyword{snapshot}
