% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVisibles.R
\name{select_visibles}
\alias{select_visibles}
\title{select_visibles function}
\usage{
select_visibles(snap, mintype = -1, maxtype = 10)
}
\arguments{
\item{snap}{dataframe obtained from reading an extended snapshot}

\item{mintype}{minimum star type to include (defaults to -1, don't touch it)}

\item{maxtype}{maximum star type to include (I chose 10 at first, try 5 or even 3)}
}
\description{
This function selects only visible star types from the snapshot
}
\examples{
select_visibles()
}
\keyword{MOCCA}
\keyword{read}
\keyword{snapshot}
