% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupROCPlot.R
\name{SetupROCPlot}
\alias{SetupROCPlot}
\title{SetupROCPlot function}
\usage{
SetupROCPlot(pdfFile, dashedidentity = TRUE)
}
\arguments{
\item{pdfFile}{File to write the pdf plot on}

\item{dashedidentity}{add a (dashed) diagonal identity line}
}
\description{
This function set up the pdf file for plotting the ROC, by setting limits
and drawing an optional diagonal line corresponding to random classification
}
\examples{
SetupROCPlot("myplot.pdf", dashedidentity=FALSE) #plot on myplot.pdf without the identity line
}
\keyword{ROC,}
\keyword{plot}
