% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncatePath.R
\name{truncate_path}
\alias{truncate_path}
\title{truncate_path function}
\usage{
truncate_path(t)
}
\arguments{
\item{t}{path to truncate}
}
\description{
This function truncates the last character off a path (to remove the final "/")
}
\examples{
truncate_path("Ciao bello!") #should return "Ciao bello" (without the "!")
}
\keyword{path}
