% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueIdentifierFromPath.R
\name{UniqueIdentifierFromPath}
\alias{UniqueIdentifierFromPath}
\title{UniqueIdentifierFromPath function}
\usage{
UniqueIdentifierFromPath(DirtyPath)
}
\arguments{
\item{DirtyPath}{path to clean up}
}
\description{
accepts a path to a simulation in the form
/data5MM/pasquato/n=700000/zini=0.005/fracb=0.1/w0=3.0/iq=1/isemi=0/ikroupa=0/rbar=120.0/rplum=25.0/kfallb=0 or in the form /data5MM/pasquato/n=700000/zini=0.005/fracb=0.1/w0=3.0/iq=1/isemi=0/ikroupa=0/rbar=120.0/rplum=25.0/kfallb=0/extsnap_1.20077167E+04.dat
the latter includes the snapshot file name
both are reduced to the first form, which is returned
}
\examples{
UniqueIdentifierFromPath("/data5MM/pasquato/n=700000/zini=0.005/fracb=0.1/w0=3.0/iq=1/isemi=0/ikroupa=0/rbar=120.0/rplum=25.0/kfallb=0 or in the form /data5MM/pasquato/n=700000/zini=0.005/fracb=0.1/w0=3.0/iq=1/isemi=0/ikroupa=0/rbar=120.0/rplum=25.0/kfallb=0/extsnap_1.20077167E+04.dat")
}
\keyword{housekeeping}
